Use more robust implementation of MIME codecs taken from the KMime library
The Codecs, Encoder and Decoder APIs as well as the actual implementations
of codecs are taken from the KMime library in kdepimlibs. This patch ports
all existing methods in KCodecs to use the new implementations, providing
better and more robust support for MIME codecs.

It also introduces KCodecs::encodeRFC2047String() method as a counterpart
to already existing KCodecs::decodeRFC2047String().

REVIEW: 121047
CHANGELOG: