Generate gperf output at build time
Summary:
Look for gperf, and use it to generate the hash-based lookup for the
entities; this replaces the static generated file in the sources,
adding a build time only dependency on gperf.

This uses the newly added FindGperf.cmake in ECM, see [1].

[1] https://phabricator.kde.org/D3830

Test Plan:
Builds fine as before, and the gperf output in the build directory
matches the removed generated source.

Reviewers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D3883