Compile benchmark tests, but don't run them automatically with ctest
Benchmarks aren't useful if no one is looking at the results (either when run
locally or on the CI); change base64benchmark so that it's always compiled, so
as not to bit-rot, but don't run it automatically when using e.g. `ctest` or
`make test`.

See https://invent.kde.org/frameworks/kio/-/merge_requests/301