deprecate QTextCodec public api
QTextCodec has been deprecated in Qt6, so also deprecate the api we have
using it.

This is not fully perfect, given we have more mappings than QTextCodec
itself, but it is the hand we have been given for now.

Apps cans till use QTextCodec by pulling in Qt5Compat, but we should not
have it on public api.