[KEncodingProberTest] Add more test cases for UTF-16
The statemachine for UTF16 (both LE and BE) fails to detect some
definitely invalid sequences - e.g. 0xdc as a high byte signals a
low surrogate, which must be preceded with a high surrogate).

On the other hand, it flags valid sequences as invalid.

As a result, UTF-16 LE and BE detection often returns the byte-swapped
result.