[KEncodingProber] Untangle HebrewProber and SBCharSetProber
The HebrewProber uses two SBCharSetProber's to get a sequence model
confidence score. But instead of handling these itself it depends
on the outer SBCSGroupProber to create and push the data to it.

The SBCharSetProber on the other hand then needs a reference (pointer)
to the parent HebrewProber to determine the correct of the two possible
charsets.

Move the two sequence models into the HebrewProber, and pass on the
input data to each one. This removes the need for the "mNameProber"
back reference in the SBCharsetProber.

Also replace the mReversed member with a template parameter, so it is
no longer branch evaluated at runtime.